refactor picklist constants to /picklist/constants.ts
