+++ packages/components/src/internal/components/domainproperties/list/constants.ts
@@ -11,6 +11,3 @@ export const SEARCH_INDEXING_TIP = 'Controls how this list is indexed for search
 export const VAR_LIST = 'VarList';
 export const INT_LIST = 'IntList';
 export const PICKLIST = 'Picklist';
-
-export const PRIVATE_PICKLIST_CATEGORY = 'PrivatePicklist';
-export const PUBLIC_PICKLIST_CATEGORY = 'PublicPicklist';
