+++ packages/components/src/internal/components/domainproperties/list/models.ts
@@ -18,7 +18,8 @@ import { Record } from 'immutable';
 import { DomainDesign, DomainField } from '../models';
 import { DOMAIN_FIELD_PRIMARY_KEY_LOCKED } from '../constants';
 
-import { INT_LIST, PICKLIST, PRIVATE_PICKLIST_CATEGORY, PUBLIC_PICKLIST_CATEGORY, VAR_LIST } from './constants';
+import { INT_LIST, PICKLIST, VAR_LIST } from './constants';
+import { PRIVATE_PICKLIST_CATEGORY, PUBLIC_PICKLIST_CATEGORY } from '../../picklist/constants';
 
 export interface AdvancedSettingsForm {
     titleColumn?: string;
