+++ packages/components/src/internal/components/picklist/constants.ts
@@ -1,0 +1,2 @@
+export const PRIVATE_PICKLIST_CATEGORY = 'PrivatePicklist';
+export const PUBLIC_PICKLIST_CATEGORY = 'PublicPicklist';
